+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_cross_macos_host_test.dart
@@ -90,4 +90,69 @@ void main() {
       }
     }
   }
+
+  const flutterMacOSLowestBestEffort = 12;
+  const flutterMacOSLowestSupported = 13;
+
+  for (final macosVersion in [
+    flutterMacOSLowestBestEffort,
+    flutterMacOSLowestSupported
+  ]) {
+    for (final linkMode in [DynamicLoadingBundled(), StaticLinking()]) {
+      test('$linkMode macos min version $macosVersion', () async {
+        const target = Architecture.arm64;
+        final tempUri = await tempDirForTest();
+        final out1Uri = tempUri.resolve('out1/');
+        await Directory.fromUri(out1Uri).create();
+        final lib1Uri = await buildLib(out1Uri, target, macosVersion, linkMode);
+
+        final otoolResult = await runProcess(
+          executable: Uri.file('otool'),
+          arguments: ['-l', lib1Uri.path],
+          logger: logger,
+        );
+        expect(otoolResult.exitCode, 0);
+        expect(otoolResult.stdout, contains('minos $macosVersion.0'));
+      });
+    }
+  }
+}
+
+Future<Uri> buildLib(
+  Uri tempUri,
+  Architecture targetArchitecture,
+  int targetMacOSVersion,
+  LinkMode linkMode,
+) async {
+  final addCUri = packageUri.resolve('test/cbuilder/testfiles/add/src/add.c');
+  const name = 'add';
+
+  final buildConfig = BuildConfig.build(
+    outputDirectory: tempUri,
+    packageName: name,
+    packageRoot: tempUri,
+    targetArchitecture: targetArchitecture,
+    targetOS: OS.macOS,
+    targetMacOSVersion: targetMacOSVersion,
+    buildMode: BuildMode.release,
+    linkModePreference: linkMode == DynamicLoadingBundled()
+        ? LinkModePreference.dynamic
+        : LinkModePreference.static,
+  );
+  final buildOutput = BuildOutput();
+
+  final cbuilder = CBuilder.library(
+    name: name,
+    assetName: name,
+    sources: [addCUri.toFilePath()],
+    dartBuildFiles: ['hook/build.dart'],
+  );
+  await cbuilder.run(
+    config: buildConfig,
+    output: buildOutput,
+    logger: logger,
+  );
+
+  final libUri = tempUri.resolve(OS.iOS.libraryFileName(name, linkMode));
+  return libUri;
 }