+++ pkgs/native_toolchain_c/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Renamed parameters in `Builder.run`.
 - Added `Language.objectiveC`.
+- Use `HookConfig.targetIosSdk` and `HookConfig.targetMacosSdk` optional
+  values, and pass them to the clang compiler.
 
 ## 0.4.2